+++ instrumentation/runtime-telemetry-jfr/library/src/main/java/io/opentelemetry/instrumentation/runtimetelemetryjfr/HandlerRegistry.java
@@ -7,9 +7,8 @@ package io.opentelemetry.instrumentation.runtimetelemetryjfr;
 
 import io.opentelemetry.api.OpenTelemetry;
 import io.opentelemetry.api.metrics.Meter;
-import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.GarbageCollection.G1GarbageCollectionHandler;
-import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.GarbageCollection.OldGarbageCollectionHandler;
-import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.GarbageCollection.YoungGarbageCollectionHandler;
+import io.opentelemetry.api.metrics.MeterBuilder;
+import io.opentelemetry.instrumentation.api.internal.EmbeddedInstrumentationProperties;
 import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.RecordedEventHandler;
 import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.ThreadGrouper;
 import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.buffer.DirectBufferStatisticsHandler;
@@ -17,17 +16,19 @@ import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.classes.Cla
 import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.container.ContainerConfigurationHandler;
 import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.cpu.ContextSwitchRateHandler;
 import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.cpu.LongLockHandler;
-import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.cpu.OverallCPULoadHandler;
+import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.cpu.OverallCpuLoadHandler;
+import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.garbagecollection.G1GarbageCollectionHandler;
+import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.garbagecollection.OldGarbageCollectionHandler;
+import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.garbagecollection.YoungGarbageCollectionHandler;
 import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.memory.CodeCacheConfigurationHandler;
 import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.memory.G1HeapSummaryHandler;
 import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.memory.MetaspaceSummaryHandler;
-import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.memory.ObjectAllocationInNewTLABHandler;
-import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.memory.ObjectAllocationOutsideTLABHandler;
+import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.memory.ObjectAllocationInNewTlabHandler;
+import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.memory.ObjectAllocationOutsideTlabHandler;
 import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.memory.ParallelHeapSummaryHandler;
 import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.network.NetworkReadHandler;
 import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.network.NetworkWriteHandler;
 import io.opentelemetry.instrumentation.runtimetelemetryjfr.internal.threads.ThreadCountHandler;
-import io.opentelemetry.instrumentation.api.internal.EmbeddedInstrumentationProperties;
 import java.io.IOException;
 import java.lang.management.GarbageCollectorMXBean;
 import java.lang.management.ManagementFactory;
@@ -36,7 +37,6 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.function.Predicate;
 import javax.annotation.Nullable;
-import io.opentelemetry.api.metrics.MeterBuilder;
 
 final class HandlerRegistry {
   private static final String SCOPE_NAME = "io.opentelemetry.instrumentation.runtimetelemetryjfr";
@@ -88,12 +88,12 @@ final class HandlerRegistry {
     ThreadGrouper grouper = new ThreadGrouper();
     List<RecordedEventHandler> basicHandlers =
         List.of(
-            new ObjectAllocationInNewTLABHandler(meter, grouper),
-            new ObjectAllocationOutsideTLABHandler(meter, grouper),
+            new ObjectAllocationInNewTlabHandler(meter, grouper),
+            new ObjectAllocationOutsideTlabHandler(meter, grouper),
             new NetworkReadHandler(meter, grouper),
             new NetworkWriteHandler(meter, grouper),
             new ContextSwitchRateHandler(meter),
-            new OverallCPULoadHandler(meter),
+            new OverallCpuLoadHandler(meter),
             new ContainerConfigurationHandler(meter),
             new LongLockHandler(meter, grouper),
             new ThreadCountHandler(meter),
