+++ instrumentation/runtime-telemetry-jfr/library/src/main/java/io/opentelemetry/instrumentation/runtimetelemetryjfr/internal/DurationUtil.java
@@ -8,7 +8,8 @@ package io.opentelemetry.instrumentation.runtimetelemetryjfr.internal;
 import java.time.Duration;
 
 /**
- * This class is internal and is hence not for public use. Its APIs are unstable and can change at any time.
+ * This class is internal and is hence not for public use. Its APIs are unstable and can change at
+ * any time.
  */
 public final class DurationUtil {
   private static final double NANOS_PER_MILLI = 1e6;
