+++ instrumentation/runtime-telemetry-jfr/library/src/main/java/io/opentelemetry/instrumentation/runtimetelemetryjfr/internal/container/ContainerConfigurationHandler.java
@@ -15,7 +15,8 @@ import java.util.List;
 import jdk.jfr.consumer.RecordedEvent;
 
 /**
- * This class is internal and is hence not for public use. Its APIs are unstable and can change at any time.
+ * This class is internal and is hence not for public use. Its APIs are unstable and can change at
+ * any time.
  */
 public final class ContainerConfigurationHandler implements RecordedEventHandler {
   private static final String METRIC_NAME = "process.runtime.jvm.cpu.limit";
