+++ instrumentation/runtime-telemetry-jfr/library/src/main/java/io/opentelemetry/instrumentation/runtimetelemetryjfr/internal/cpu/ContextSwitchRateHandler.java
@@ -17,7 +17,8 @@ import java.util.Optional;
 import jdk.jfr.consumer.RecordedEvent;
 
 /**
- * This class is internal and is hence not for public use. Its APIs are unstable and can change at any time.
+ * This class is internal and is hence not for public use. Its APIs are unstable and can change at
+ * any time.
  */
 public final class ContextSwitchRateHandler implements RecordedEventHandler {
   private static final String METRIC_NAME = "process.runtime.jvm.cpu.context_switch";
