+++ instrumentation/runtime-telemetry-jfr/library/src/main/java/io/opentelemetry/instrumentation/runtimetelemetryjfr/internal/cpu/LongLockHandler.java
@@ -21,7 +21,8 @@ import java.util.function.Consumer;
 import jdk.jfr.consumer.RecordedEvent;
 
 /**
- * This class is internal and is hence not for public use. Its APIs are unstable and can change at any time.
+ * This class is internal and is hence not for public use. Its APIs are unstable and can change at
+ * any time.
  */
 public final class LongLockHandler extends AbstractThreadDispatchingHandler {
   private static final String METRIC_NAME = "process.runtime.jvm.cpu.longlock";
