+++ instrumentation/runtime-telemetry-jfr/library/src/main/java/io/opentelemetry/instrumentation/runtimetelemetryjfr/internal/memory/ObjectAllocationInNewTlabHandler.java
@@ -21,17 +21,18 @@ import java.util.function.Consumer;
 import jdk.jfr.consumer.RecordedEvent;
 
 /**
- * This class handles all non-TLAB allocation JFR events, and delegates them to the actual
- * per-thread aggregators
+ * This class handles TLAB allocation JFR events, and delegates them to the actual per-thread
+ * aggregators
  *
- * This class is internal and is hence not for public use. Its APIs are unstable and can change at any time.
+ * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
+ * at any time.
  */
-public final class ObjectAllocationOutsideTLABHandler extends AbstractThreadDispatchingHandler {
-  private static final String EVENT_NAME = "jdk.ObjectAllocationOutsideTLAB";
+public final class ObjectAllocationInNewTlabHandler extends AbstractThreadDispatchingHandler {
+  private static final String EVENT_NAME = "jdk.ObjectAllocationInNewTLAB";
 
   private final LongHistogram histogram;
 
-  public ObjectAllocationOutsideTLABHandler(Meter meter, ThreadGrouper grouper) {
+  public ObjectAllocationInNewTlabHandler(Meter meter, ThreadGrouper grouper) {
     super(grouper);
     histogram =
         meter
@@ -54,28 +55,28 @@ public final class ObjectAllocationOutsideTLABHandler extends AbstractThreadDisp
 
   @Override
   public Consumer<RecordedEvent> createPerThreadSummarizer(String threadName) {
-    return new PerThreadObjectAllocationOutsideTLABHandler(histogram, threadName);
+    return new PerThreadObjectAllocationInNewTlabHandler(histogram, threadName);
   }
 
   @Override
   public void close() {}
 
-  /** This class aggregates all non-TLAB allocation JFR events for a single thread */
-  private static class PerThreadObjectAllocationOutsideTLABHandler
+  /** This class aggregates all TLAB allocation JFR events for a single thread */
+  private static class PerThreadObjectAllocationInNewTlabHandler
       implements Consumer<RecordedEvent> {
-    private static final String ALLOCATION_SIZE = "allocationSize";
+    private static final String TLAB_SIZE = "tlabSize";
 
     private final LongHistogram histogram;
     private final Attributes attributes;
 
-    public PerThreadObjectAllocationOutsideTLABHandler(LongHistogram histogram, String threadName) {
+    public PerThreadObjectAllocationInNewTlabHandler(LongHistogram histogram, String threadName) {
       this.histogram = histogram;
-      this.attributes = Attributes.of(ATTR_THREAD_NAME, threadName, ATTR_ARENA_NAME, "Main");
+      this.attributes = Attributes.of(ATTR_THREAD_NAME, threadName, ATTR_ARENA_NAME, "TLAB");
     }
 
     @Override
     public void accept(RecordedEvent ev) {
-      histogram.record(ev.getLong(ALLOCATION_SIZE), attributes);
+      histogram.record(ev.getLong(TLAB_SIZE), attributes);
       // Probably too high a cardinality
       // ev.getClass("objectClass").getName();
     }
