+++ instrumentation/runtime-telemetry-jfr/library/src/main/java/io/opentelemetry/instrumentation/runtimetelemetryjfr/internal/network/NetworkReadHandler.java
@@ -27,7 +27,8 @@ import java.util.function.Consumer;
 import jdk.jfr.consumer.RecordedEvent;
 
 /**
- * This class is internal and is hence not for public use. Its APIs are unstable and can change at any time.
+ * This class is internal and is hence not for public use. Its APIs are unstable and can change at
+ * any time.
  */
 public final class NetworkReadHandler extends AbstractThreadDispatchingHandler {
   private static final String EVENT_NAME = "jdk.SocketRead";
