+++ instrumentation/runtime-telemetry-jfr/library/src/main/java/io/opentelemetry/instrumentation/runtimetelemetryjfr/internal/network/NetworkWriteHandler.java
@@ -45,7 +45,8 @@ import jdk.jfr.consumer.RecordedEvent;
 // }
 
 /**
- * This class is internal and is hence not for public use. Its APIs are unstable and can change at any time.
+ * This class is internal and is hence not for public use. Its APIs are unstable and can change at
+ * any time.
  */
 public final class NetworkWriteHandler extends AbstractThreadDispatchingHandler {
   private static final String EVENT_NAME = "jdk.SocketWrite";
