+++ instrumentation/runtime-telemetry-jfr/library/src/main/java/io/opentelemetry/instrumentation/runtimetelemetryjfr/internal/threads/ThreadCountHandler.java
@@ -19,7 +19,8 @@ import java.util.Optional;
 import jdk.jfr.consumer.RecordedEvent;
 
 /**
- * This class is internal and is hence not for public use. Its APIs are unstable and can change at any time.
+ * This class is internal and is hence not for public use. Its APIs are unstable and can change at
+ * any time.
  */
 public final class ThreadCountHandler implements RecordedEventHandler {
   private static final String METRIC_NAME = "process.runtime.jvm.threads.count";
